pyproject.toml
[project]
name = "document-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.3",
    "fastapi>=0.121.0",
    "numpy>=2.3.4",
    "openai>=2.7.2",
    "pdfplumber>=0.11.7",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=0.2.5.1",
    "uvicorn[standard]>=0.38.0",
]
